requirements.txt
numpy
python-dotenv
requests
python-telegram-bot